# generated-output/wildfly-swarm/openapi-jaxrs-example.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Generated API",
    "version" : "1.0"
  },
  "paths" : {
...
    "/organizations" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewOrganizationBean"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrganizationBean"
                }
              }
            }
          }
        }
      }
    },
...
    "/search/apiCatalog/entries" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SearchCriteriaBean"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "beans" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "definitionType" : {
                            "enum" : [ "External", "None", "RAML", "SwaggerJSON", "SwaggerYAML", "WADL", "WSDL" ],
                            "type" : "string"
                          },
                          "definitionUrl" : {
                            "type" : "string"
                          },
                          "description" : {
                            "type" : "string"
                          },
                          "endpoint" : {
                            "type" : "string"
                          },
                          "endpointType" : {
                            "enum" : [ "rest", "soap" ],
                            "type" : "string"
                          },
                          "icon" : {
                            "type" : "string"
                          },
                          "id" : {
                            "type" : "string"
                          },
                          "internal" : {
                            "type" : "boolean"
                          },
                          "name" : {
                            "type" : "string"
                          },
                          "namespace" : {
                            "type" : "string"
                          },
                          "routeDefinitionUrl" : {
                            "type" : "string"
                          },
                          "routeEndpoint" : {
                            "type" : "string"
                          },
                          "tags" : {
                            "type" : "array",
                            "items" : {
                              "type" : "string"
                            }
                          }
                        }
                      }
                    },
                    "totalSize" : {
                      "format" : "int32",
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/search/apis" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SearchCriteriaBean"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "beans" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "createdOn" : {
                            "format" : "date",
                            "type" : "string"
                          },
                          "description" : {
                            "type" : "string"
                          },
                          "id" : {
                            "type" : "string"
                          },
                          "name" : {
                            "type" : "string"
                          },
                          "organizationId" : {
                            "type" : "string"
                          },
                          "organizationName" : {
                            "type" : "string"
                          }
                        }
                      }
                    },
                    "totalSize" : {
                      "format" : "int32",
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search/clients" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SearchCriteriaBean"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "beans" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "description" : {
                            "type" : "string"
                          },
                          "id" : {
                            "type" : "string"
                          },
                          "name" : {
                            "type" : "string"
                          },
                          "numContracts" : {
                            "format" : "int32",
                            "type" : "integer"
                          },
                          "organizationId" : {
                            "type" : "string"
                          },
                          "organizationName" : {
                            "type" : "string"
                          }
                        }
                      }
                    },
                    "totalSize" : {
                      "format" : "int32",
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/search/organizations" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SearchCriteriaBean"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "properties" : {
                    "beans" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "description" : {
                            "type" : "string"
                          },
                          "id" : {
                            "type" : "string"
                          },
                          "name" : {
                            "type" : "string"
                          },
                          "numApis" : {
                            "format" : "int32",
                            "type" : "integer"
                          },
                          "numClients" : {
                            "format" : "int32",
                            "type" : "integer"
                          },
                          "numMembers" : {
                            "format" : "int32",
                            "type" : "integer"
                          }
                        }
                      }
                    },
                    "totalSize" : {
                      "format" : "int32",
                      "type" : "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "NewOrganizationBean" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "OrganizationBean" : {
        "properties" : {
          "apis" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "createdBy" : {
                  "type" : "string"
                },
                "createdOn" : {
                  "format" : "date",
                  "type" : "string"
                },
                "description" : {
                  "type" : "string"
                },
                "id" : {
                  "type" : "string"
                },
                "name" : {
                  "type" : "string"
                },
                "numPublished" : {
                  "format" : "int32",
                  "type" : "integer"
                },
                "organization" : {
                  "description" : "Cyclic reference to org.example.microprofile.api.beans.orgs.OrganizationBean",
                  "type" : "object"
                },
                "versions" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "properties" : {
                      "api" : {
                        "description" : "Cyclic reference to org.example.microprofile.api.beans.apis.ApiBean",
                        "type" : "object"
                      },
                      "createdBy" : {
                        "type" : "string"
                      },
                      "createdOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "definition" : {
                        "type" : "object",
                        "properties" : {
                          "apiVersion" : {
                            "description" : "Cyclic reference to org.example.microprofile.api.beans.apis.ApiVersionBean",
                            "type" : "object"
                          },
                          "data" : {
                            "type" : "array",
                            "items" : {
                              "format" : "byte",
                              "type" : "string"
                            }
                          },
                          "id" : {
                            "format" : "int64",
                            "type" : "integer"
                          }
                        }
                      },
                      "definitionType" : {
                        "enum" : [ "External", "None", "RAML", "SwaggerJSON", "SwaggerYAML", "WADL", "WSDL" ],
                        "type" : "string"
                      },
                      "endpoint" : {
                        "type" : "string"
                      },
                      "endpointContentType" : {
                        "enum" : [ "json", "xml" ],
                        "type" : "string"
                      },
                      "endpointProperties" : {
                        "type" : "object",
                        "additionalProperties" : {
                          "type" : "string"
                        }
                      },
                      "endpointType" : {
                        "enum" : [ "rest", "soap" ],
                        "type" : "string"
                      },
                      "gateways" : {
                        "type" : "array",
                        "items" : {
                          "type" : "object",
                          "properties" : {
                            "gatewayId" : {
                              "type" : "string"
                            }
                          }
                        }
                      },
                      "id" : {
                        "format" : "int64",
                        "type" : "integer"
                      },
                      "modifiedBy" : {
                        "type" : "string"
                      },
                      "modifiedOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "parsePayload" : {
                        "type" : "boolean"
                      },
                      "plans" : {
                        "type" : "array",
                        "items" : {
                          "type" : "object",
                          "properties" : {
                            "planId" : {
                              "type" : "string"
                            },
                            "version" : {
                              "type" : "string"
                            }
                          }
                        }
                      },
                      "publicAPI" : {
                        "type" : "boolean"
                      },
                      "publishedOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "retiredOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "status" : {
                        "enum" : [ "Created", "Published", "Ready", "Retired" ],
                        "type" : "string"
                      },
                      "version" : {
                        "type" : "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "clients" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "createdBy" : {
                  "type" : "string"
                },
                "createdOn" : {
                  "format" : "date",
                  "type" : "string"
                },
                "description" : {
                  "type" : "string"
                },
                "id" : {
                  "type" : "string"
                },
                "name" : {
                  "type" : "string"
                },
                "organization" : {
                  "description" : "Cyclic reference to org.example.microprofile.api.beans.orgs.OrganizationBean",
                  "type" : "object"
                },
                "versions" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "properties" : {
                      "apikey" : {
                        "type" : "string"
                      },
                      "client" : {
                        "description" : "Cyclic reference to org.example.microprofile.api.beans.clients.ClientBean",
                        "type" : "object"
                      },
                      "createdBy" : {
                        "type" : "string"
                      },
                      "createdOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "id" : {
                        "format" : "int64",
                        "type" : "integer"
                      },
                      "modifiedBy" : {
                        "type" : "string"
                      },
                      "modifiedOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "publishedOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "retiredOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "status" : {
                        "enum" : [ "Created", "Ready", "Registered", "Retired" ],
                        "type" : "string"
                      },
                      "version" : {
                        "type" : "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "createdBy" : {
            "type" : "string"
          },
          "createdOn" : {
            "format" : "date",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "modifiedBy" : {
            "type" : "string"
          },
          "modifiedOn" : {
            "format" : "date",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "plans" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "createdBy" : {
                  "type" : "string"
                },
                "createdOn" : {
                  "format" : "date",
                  "type" : "string"
                },
                "description" : {
                  "type" : "string"
                },
                "id" : {
                  "type" : "string"
                },
                "name" : {
                  "type" : "string"
                },
                "organization" : {
                  "description" : "Cyclic reference to org.example.microprofile.api.beans.orgs.OrganizationBean",
                  "type" : "object"
                },
                "versions" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "properties" : {
                      "createdBy" : {
                        "type" : "string"
                      },
                      "createdOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "id" : {
                        "format" : "int64",
                        "type" : "integer"
                      },
                      "lockedOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "modifiedBy" : {
                        "type" : "string"
                      },
                      "modifiedOn" : {
                        "format" : "date",
                        "type" : "string"
                      },
                      "plan" : {
                        "description" : "Cyclic reference to org.example.microprofile.api.beans.plans.PlanBean",
                        "type" : "object"
                      },
                      "status" : {
                        "enum" : [ "Created", "Locked", "Ready" ],
                        "type" : "string"
                      },
                      "version" : {
                        "type" : "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "SearchCriteriaBean" : {
        "properties" : {
          "filters" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "name" : {
                  "type" : "string"
                },
                "operator" : {
                  "enum" : [ "bool_eq", "eq", "gt", "gte", "like", "lt", "lte", "neq" ],
                  "type" : "string"
                },
                "value" : {
                  "type" : "string"
                }
              }
            }
          },
          "orderBy" : {
            "type" : "object",
            "properties" : {
              "ascending" : {
                "type" : "boolean"
              },
              "name" : {
                "type" : "string"
              }
            }
          },
          "paging" : {
            "type" : "object",
            "properties" : {
              "page" : {
                "format" : "int32",
                "type" : "integer"
              },
              "pageSize" : {
                "format" : "int32",
                "type" : "integer"
              }
            }
          }
        }
      },
...
    }
  }
}